+++ MatrixSDK/Crypto/MXCrypto.h
@@ -292,6 +292,15 @@ FOUNDATION_EXPORT NSString *const kMXCryptoRoomKeyRequestCancellationNotificatio
 
 
 /**
+ Retrieve the known devices for a user.
+
+ @param userId The user id.
+ @return A map from device id to 'MXDevice' object for the device or nil if we
+         haven't managed to get a list of devices for this user yet.
+ */
+- (NSDictionary<NSString*, MXDeviceInfo*>*)devicesForUser:(NSString*)userId;
+
+/**
  Get the stored information about a device.
 
  @param deviceId The device.