+++ CHANGES.rst
@@ -6,6 +6,7 @@ Improvements:
  * Aggregations: Implement m.reference aggregations, aka thread ([MSC1849](https://github.com/matrix-org/matrix-doc/blob/matthew/msc1849/proposals/1849-aggregations.md)).
  * Cross-Signing: Add a new module, MXCrossSigning, to handle device cross-signing (vector-im/riot-ios/issues/2890).
  * MXRoom: Add a method to retrieve trusted members count in an encrypted room.
+ * MXCrypto: Expose devicesForUser.
 
 Bug fix:
  * MXEventType: Fix Swift refinement.